UPSTREAM: clk: spacemit: prepare common ccu header
In order to prepare adding clock driver for new K3 SoC, extract generic
code to a separate common ccu header file, so they are not defined
in K1 SoC-specific file, and then can be shared by all clock drivers.

Link: https://lore.kernel.org/r/20260108-06-k1-clk-common-v4-1-badf635993d3@gentoo.org
(cherry picked from commit 2b7a02c322922a37cc5fc15d055b794cc2193062)
